POCOR-3853 - Clean up for infrastructure items. Fixed institution action button.

+++ plugins/Institution/src/Controller/InstitutionsController.php
@@ -25,7 +25,7 @@ class InstitutionsController extends AppController
     public function initialize()
     {
         parent::initialize();
-        $this->ControllerAction->model('Institution.Institutions', [], ['deleteStrategy' => 'restrict']);
+        // $this->ControllerAction->model('Institution.Institutions', [], ['deleteStrategy' => 'restrict']);
         $this->ControllerAction->models = [
             'Attachments'       => ['className' => 'Institution.InstitutionAttachments'],
             'History'           => ['className' => 'Institution.InstitutionActivities', 'actions' => ['search', 'index']],
@@ -73,6 +73,11 @@ class InstitutionsController extends AppController
     }
 
     // CAv4
+    public function Institutions()
+    {
+        $this->ControllerAction->process(['alias' => __FUNCTION__, 'className' => 'Institution.Institutions']);
+    }
+
     public function Positions()
     {
         $this->ControllerAction->process(['alias' => __FUNCTION__, 'className' => 'Institution.InstitutionPositions']);
@@ -258,7 +263,7 @@ class InstitutionsController extends AppController
 
         if (!$this->AccessControl->isAdmin()) {
             $userId = $this->Auth->user('id');
-            $roles = $this->Institutions->getInstitutionRoles($userId, $institutionId);
+            $roles = TableRegistry::get('Institution.Institutions')->getInstitutionRoles($userId, $institutionId);
         }
 
         $this->set('_roles', $roles);
@@ -305,7 +310,7 @@ class InstitutionsController extends AppController
             $this->Navigation->addCrumb($crumbTitle, $indexUrl);
             if (!$this->AccessControl->isAdmin() && $institutionId) {
                 $userId = $this->Auth->user('id');
-                $roles = $this->Institutions->getInstitutionRoles($userId, $institutionId);
+                $roles = TableRegistry::get('Institution.Institutions')->getInstitutionRoles($userId, $institutionId);
                 $AccessControl = $this->AccessControl;
                 $action = 'edit';
                 if (!$AccessControl->check(['Institutions', 'StudentCompetencies', $action], $roles)) {
@@ -346,7 +351,7 @@ class InstitutionsController extends AppController
             $institutionId = !empty($this->request->param('institutionId')) ? $this->ControllerAction->paramsDecode($this->request->param('institutionId'))['id'] : $session->read('Institution.Institutions.id');
             if (!$this->AccessControl->isAdmin() && $institutionId) {
                 $userId = $this->Auth->user('id');
-                $roles = $this->Institutions->getInstitutionRoles($userId, $institutionId);
+                $roles = TableRegistry::get('Institution.Institutions')->getInstitutionRoles($userId, $institutionId);
                 $AccessControl = $this->AccessControl;
                 $action = 'edit';
                 if (!$AccessControl->check(['Institutions', 'AllClasses', $action], $roles)) {
@@ -411,7 +416,7 @@ class InstitutionsController extends AppController
             $institutionId = !empty($this->request->param('institutionId')) ? $this->ControllerAction->paramsDecode($this->request->param('institutionId'))['id'] : $session->read('Institution.Institutions.id');
             if (!$this->AccessControl->isAdmin() && $institutionId) {
                 $userId = $this->Auth->user('id');
-                $roles = $this->Institutions->getInstitutionRoles($userId, $institutionId);
+                $roles = TableRegistry::get('Institution.Institutions')->getInstitutionRoles($userId, $institutionId);
                 $AccessControl = $this->AccessControl;
                 $action = 'edit';
                 if (!$AccessControl->check(['Institutions', 'AllSubjects', $action], $roles)) {
@@ -463,7 +468,7 @@ class InstitutionsController extends AppController
             if (!$this->AccessControl->isAdmin() && $session->check('Institution.Institutions.id')) {
                 $userId = $this->Auth->user('id');
                 $institutionId = $session->read('Institution.Institutions.id');
-                $roles = $this->Institutions->getInstitutionRoles($userId, $institutionId);
+                $roles = TableRegistry::get('Institution.Institutions')->getInstitutionRoles($userId, $institutionId);
             }
             $this->set('ngController', 'InstitutionsStudentsCtrl as InstitutionStudentController');
             $this->set('_createNewStudent', $this->AccessControl->check(['Institutions', 'getUniqueOpenemisId'], $roles));
@@ -489,7 +494,7 @@ class InstitutionsController extends AppController
             if (!$this->AccessControl->isAdmin() && $session->check('Institution.Institutions.id')) {
                 $userId = $this->Auth->user('id');
                 $institutionId = $session->read('Institution.Institutions.id');
-                $roles = $this->Institutions->getInstitutionRoles($userId, $institutionId);
+                $roles = TableRegistry::get('Institution.Institutions')->getInstitutionRoles($userId, $institutionId);
             }
             $this->set('ngController', 'InstitutionsStaffCtrl as InstitutionStaffController');
             $this->set('_createNewStaff', $this->AccessControl->check(['Institutions', 'getUniqueOpenemisId'], $roles));
@@ -564,7 +569,7 @@ class InstitutionsController extends AppController
     {
         parent::beforeFilter($event);
         $session = $this->request->session();
-        $this->Navigation->addCrumb('Institutions', ['plugin' => 'Institution', 'controller' => 'Institutions', 'action' => 'index']);
+        $this->Navigation->addCrumb('Institutions', ['plugin' => 'Institution', 'controller' => 'Institutions', 'action' => 'Institutions', 'index']);
         $action = $this->request->params['action'];
         $header = __('Institutions');
 
@@ -587,8 +592,7 @@ class InstitutionsController extends AppController
             }
             $session->write('Institution.Institutions.id', $query['institution_id']);
         }
-
-        if ($action == 'index') {
+        if ($action == 'Institutions' && isset($this->request->pass[0]) && $this->request->pass[0] == 'index') {
             $session->delete('Institution.Institutions');
         } elseif ($action == 'StudentUser') {
             $session->write('Student.Students.id', $this->ControllerAction->paramsDecode($this->request->pass[1])['id']);
@@ -596,7 +600,10 @@ class InstitutionsController extends AppController
             $session->write('Staff.Staff.id', $this->ControllerAction->paramsDecode($this->request->pass[1])['id']);
         }
 
-        if (($session->check('Institution.Institutions.id') || $this->request->param('institutionId')) || in_array($action, ['view', 'edit', 'dashboard'])) {
+        if (($session->check('Institution.Institutions.id')
+            || $this->request->param('institutionId'))
+            || $action == 'dashboard'
+            || ($action == 'Institutions' && isset($this->request->pass[0]) && in_array($this->request->pass[0], ['view', 'edit']))) {
             $id = 0;
             if (isset($this->request->pass[0]) && (in_array($action, ['view', 'edit', 'dashboard']))) {
                 $id = $this->request->pass[0];
@@ -615,7 +622,7 @@ class InstitutionsController extends AppController
                 $id = $session->read('Institution.Institutions.id');
             }
             if (!empty($id)) {
-                $this->activeObj = $this->Institutions->get($id);
+                $this->activeObj = TableRegistry::get('Institution.Institutions')->get($id);
                 $name = $this->activeObj->name;
                 $session->write('Institution.Institutions.name', $name);
                 if ($action == 'view') {
@@ -627,7 +634,7 @@ class InstitutionsController extends AppController
                 }
                 $this->Navigation->addCrumb($name, ['plugin' => 'Institution', 'controller' => 'Institutions', 'action' => 'dashboard', 'institutionId' => $this->ControllerAction->paramsEncode(['id' => $id]), $this->ControllerAction->paramsEncode(['id' => $id])]);
             } else {
-                return $this->redirect(['plugin' => 'Institution', 'controller' => 'Institutions', 'action' => 'index']);
+                return $this->redirect(['plugin' => 'Institution', 'controller' => 'Institutions', 'action' => 'Institutions', 'index']);
             }
         }
         $this->set('contentHeader', $header);
@@ -850,10 +857,10 @@ class InstitutionsController extends AppController
                 $this->Navigation->addCrumb($model->getHeader($model->alias()));
                 $header = __('Institutions') . ' - ' . $model->getHeader($model->alias());
                 $this->set('contentHeader', $header);
-            } else {
+            } elseif ($this->request->param('action') != 'Institutions') {
                 $this->Alert->warning('general.notExists');
                 $event->stopPropagation();
-                return $this->redirect(['plugin' => 'Institution', 'controller' => 'Institutions', 'action' => 'index']);
+                return $this->redirect(['plugin' => 'Institution', 'controller' => 'Institutions', 'action' => 'Institutions', 'index']);
             }
         }
     }
@@ -889,16 +896,16 @@ class InstitutionsController extends AppController
 
     public function excel($id = 0)
     {
-        $this->Institutions->excel($id);
+        TableRegistry::get('Institution.Institutions')->excel($id);
         $this->autoRender = false;
     }
 
     public function dashboard($id)
     {
         $id = $this->ControllerAction->paramsDecode($id)['id'];
-        $this->ControllerAction->model->action = $this->request->action;
+        // $this->ControllerAction->model->action = $this->request->action;
 
-        $Institutions = $this->Institutions;
+        $Institutions = TableRegistry::get('Institution.Institutions');
         $classification = $Institutions->get($id)->classification;
         $AcademicPeriods = TableRegistry::get('AcademicPeriod.AcademicPeriods');
         $currentPeriod = $AcademicPeriods->getCurrent();
